+++ CHANGELOG.md
@@ -5,11 +5,11 @@
   config struct that defines them.
 * An `InstaceID == '-'` or `nodeId:"-"` will ignore this field from the graph.
 * Graph configs does not need to implement `ConnectedConfig` interface if their properties define the
-  `sendsTo` configuration.
+  `sendTo` configuration.
 * Graph builder returns error if nodes remain unconnected
 * Graph builder returns error if a node sends data to itself
 * A node config implementing `Enabler` interface can allow users disabling nodes without requiring pointers.
-* Disabled or nil nodes can forward the received data by using the `fwdTo` annotation.
+* Disabled or nil nodes can forward the received data by using the `forwardTo` annotation.
 
 ## Breaking changes
 * High-level graph API add contexts to the `graph.Run(context.Context)` library, and